Website: 	Added support for: 		Loading multiple languages & holding them in memory 		Loading categories & holding them in memory. 	Updated many files to comply with changes.
DataAdministration:
	Added support for:
		Adding & Editing a product.

+++ DomainModel/Entities/ProductBrand.cs
@@ -6,7 +6,7 @@ namespace DomainModel.Entities
 {
     public class ProductBrand
     {
-        public int BrandId { get; set; }
+        public Int64 BrandId { get; set; }
         public string BrandName { get; set; }
         public string ResourceUrl { get; set; } // Images, Intro videos, etc.
     }
